analytics-platform/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "analytics-platform"
version = "0.1.0"
description = "Local Analytical Data Platform with Parquet, DuckDB, and PostgreSQL"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["analytics", "duckdb", "parquet", "postgresql", "data-engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "duckdb>=1.0.0",
    "pyarrow>=15.0.0",
    "pandas>=2.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
postgres = [
    "psycopg2-binary>=2.9.0",
]
polars = [
    "polars>=0.20.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "analytics-platform[postgres,polars,dev]",
]

[project.scripts]
analytics = "src.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/analytics-platform"
Documentation = "https://github.com/yourusername/analytics-platform#readme"
Repository = "https://github.com/yourusername/analytics-platform"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
